Treaning other/python/8.2.django-angular-trimmed/8.2.django-angular-trimmed/regex-core/requirements.txt
coverage==5.5
django==3.2.3
django-cors-headers==3.7.0
django-extensions==3.1.3
djangorestframework==3.12.2
pycodestyle==2.7.0
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
pytest-xdist==2.2.1
